5杂项、副防具

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/rings/RingOfElements.java
@@ -41,13 +41,13 @@ public class RingOfElements extends Ring {
 	public String statsInfo() {
 		if (已鉴定()){
 			String info = Messages.get(this, "stats",
-									   kw2(100f*(1f-Math.pow(0.825f,soloBuffedBonus()))),
-									   kw2(100f*(1f-Math.pow(0.825f,soloBuffedBonus())))
+									   kw2(100f*(float)(1f-Math.pow(0.825f,soloBuffedBonus()))),
+									   kw2(100f*(float)(1f-Math.pow(0.825f,soloBuffedBonus())))
 									  );
 			if (isEquipped(Dungeon.hero) && soloBuffedBonus() != combinedBuffedBonus(Dungeon.hero)){
 				info += "\n\n" + Messages.get(this, "combined_stats",
-											  kw2(100f*(1f-Math.pow(0.825f,combinedBuffedBonus(Dungeon.hero)))),
-												  kw2(00f*(1f-Math.pow(0.825f,combinedBuffedBonus(Dungeon.hero))))
+											  kw2(100f*(float)(1f-Math.pow(0.825f,combinedBuffedBonus(Dungeon.hero)))),
+												  kw2(00f*(float)(1f-Math.pow(0.825f,combinedBuffedBonus(Dungeon.hero))))
 											 );
 			}
 			return info;
